Fix failing color test.
Entries in the tree.colors hash are assumed to map to hex values. Since
transparent was mapped to `rgba(0,0,0,0)` it went crazy. We could tweak
the code to understand keywords mapped to rgba() or rgb(), but we only
need it for transparent, which works just fine as a keyword, and which
happens to be fewer bytes than rgba(0,0,0,0). So just don't bother
remapping it at all.