Adjust relative URLs in imported stylesheets.
This is almost certainly wrong, but it seems to be working for me...

When processing an imported stylesheet outside of a browser, adjust all
relative URLs so that they become relative to the output file even though
they're specified relative to the input file. This is particularly important
because it allows us to use less in a browser during development and compiled
stylesheets in production while still storing the .less files in an arbitrary
directory structure.